# docs/swagger/gopher-stamp-crud.yml
openapi: 3.0.0
info:
  title: Gopher Stamp Rally User API
  description: API for managing users in the Gopher Stamp Rally application
  version: 1.0.0
servers:
  - url: http://localhost:8080

paths:
  # User endpoints
  /users:
    get:
      summary: ユーザー一覧取得
      description: 全てのユーザーを取得する
      operationId: listUsers
      tags:
        - Users
      responses:
        '200':
          description: ユーザー一覧の取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              examples:
                sampleUsers:
                  summary: モック用の参加者一覧
                  value:
                    - id: 101
                      name: 田中太郎
                      twitter_id: tanaka_taro
                      favorite_go_feature: goroutineによる並行処理
                      icon: https://go.dev/images/gophers/ladder.svg
                    - id: 102
                      name: Gopher花子
                      twitter_id: gopher_hanako
                      favorite_go_feature: Go初心者向けセッション
                      icon: https://go.dev/images/gophers/ladder-step.svg
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: ユーザー作成
      description: 新しいユーザーを作成する
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{id}:
    get:
      summary: ユーザー詳細取得
      description: 指定されたIDのユーザーを取得する
      operationId: getUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: ユーザーID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ユーザー詳細の取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: ユーザー更新
      description: 指定されたIDのユーザーを更新する
      operationId: updateUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: ユーザーID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: ユーザー更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Stamp endpoints
  /stamps:
    get:
      summary: スタンプ一覧取得
      description: 全てのスタンプマスターデータを取得する
      operationId: listStamps
      tags:
        - Stamps
      parameters:
        - name: limit
          in: query
          description: 取得する件数の上限
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: スキップする件数
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: スタンプ一覧の取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  stamps:
                    type: array
                    items:
                      $ref: '#/components/schemas/Stamp'
                  total:
                    type: integer
                    description: 総件数
              examples:
                sampleSingleStamp:
                  summary: モック用に1件のみ返すサンプル
                  value:
                    stamps:
                      - id: 1
                        name: "Go基礎セッション"
                    total: 1
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: スタンプ作成
      description: 新しいスタンプマスターデータを作成する
      operationId: createStamp
      tags:
        - Stamps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StampCreateRequest'
      responses:
        '201':
          description: スタンプ作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stamp'
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /stamps/{id}:
    get:
      summary: スタンプ詳細取得
      description: 指定されたIDのスタンプを取得する
      operationId: getStamp
      tags:
        - Stamps
      parameters:
        - name: id
          in: path
          required: true
          description: スタンプID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: スタンプ詳細の取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stamp'
              example:
                id: 1
                name: "Go基礎セッション"
        '404':
          description: スタンプが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: スタンプ更新
      description: 指定されたIDのスタンプを更新する
      operationId: updateStamp
      tags:
        - Stamps
      parameters:
        - name: id
          in: path
          required: true
          description: スタンプID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StampUpdateRequest'
      responses:
        '200':
          description: スタンプ更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stamp'
        '400':
          description: リクエストが不正（nameは必須）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "INVALID_REQUEST"
                message: "name must be provided"
                details: "name must be provided"
        '404':
          description: スタンプが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: スタンプ削除
      description: 指定されたIDのスタンプを削除する
      operationId: deleteStamp
      tags:
        - Stamps
      parameters:
        - name: id
          in: path
          required: true
          description: スタンプID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: スタンプ削除成功
        '404':
          description: スタンプが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /stamps/{id}/otp:
    get:
      tags:
        - Stamps
      summary: スタンプのOTPを取得
      description: 指定されたIDのスタンプの現在のOTPを取得する（管理者用）。
      operationId: getStampOTP
      parameters:
        - name: id
          in: path
          description: スタンプID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  otp:
                    type: string
                    example: "1234"
        '404':
          description: スタンプが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # User Stamp endpoints (ユーザーがスタンプを取得)
  /users/{id}/stamps:
    get:
      summary: ユーザーの取得済みスタンプ一覧取得
      description: 指定されたユーザーが取得したスタンプの一覧を取得する
      operationId: listUserStamps
      tags:
        - UserStamps
      parameters:
        - name: id
          in: path
          required: true
          description: ユーザーID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ユーザーの取得済みスタンプ一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  stamps:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserStamp'
              examples:
                user-101:
                  summary: ユーザー101は0件取得済み
                  value:
                    stamps: []
                user-102:
                  summary: ユーザー102は1件取得済み（stamp_id = 1）
                  value:
                    stamps:
                      - user_id: 102
                        stamp_id: 1
                        acquired_at: "2023-01-01T00:00:00Z"
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: ユーザーがスタンプを取得
      description: ユーザーが新しいスタンプを取得する
      operationId: acquireStamp
      tags:
        - UserStamps
      parameters:
        - name: id
          in: path
          required: true
          description: ユーザーID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcquireStampRequest'
      responses:
        '201':
          description: スタンプ取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStamp'
              example:
                user_id: 1
                stamp_id: 1
                acquired_at: "2023-01-01T00:00:00Z"
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: ユーザーまたはスタンプが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 既に取得済みのスタンプ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    User:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          description: ユーザーID
          example: 1
        name:
          type: string
          description: ユーザー名
          example: "田中太郎"
          maxLength: 100
        twitter_id:
          type: string
          description: TwitterID
          example: "tanaka_taro"
          maxLength: 50
        favorite_go_feature:
          type: string
          description: 好きなGoの特徴
          example: "goroutineによる並行処理"
          maxLength: 500
        icon:
          type: string
          description: アイコン画像URL
          example: "https://example.com/icons/gopher.png"
          maxLength: 500
        created_at:
          type: string
          format: date-time
          description: 作成日時
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: 更新日時
          example: "2023-01-01T00:00:00Z"

    UserCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: ユーザー名
          example: "田中太郎"
          maxLength: 100
        twitter_id:
          type: string
          description: TwitterID
          example: "tanaka_taro"
          maxLength: 50
        favorite_go_feature:
          type: string
          description: 好きなGoの特徴
          example: "goroutineによる並行処理"
          maxLength: 500
        icon:
          type: string
          description: アイコン画像URL
          example: "https://example.com/icons/gopher.png"
          maxLength: 500

    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: ユーザー名
          example: "田中太郎"
          maxLength: 100
        twitter_id:
          type: string
          description: TwitterID
          example: "tanaka_taro"
          maxLength: 50
        favorite_go_feature:
          type: string
          description: 好きなGoの特徴
          example: "goroutineによる並行処理"
          maxLength: 500
        icon:
          type: string
          description: アイコン画像URL
          example: "https://example.com/icons/gopher.png"
          maxLength: 500

    UserDetail:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            acquired_stamps:
              type: array
              description: 取得済みスタンプの一覧
              items:
                $ref: '#/components/schemas/UserStamp'

    Stamp:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          description: スタンプID
          example: 1
        name:
          type: string
          description: スタンプ名
          example: "Gopher基礎"
          maxLength: 100

    StampCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: スタンプ名
          example: "Gopher基礎"
          maxLength: 100

    StampUpdateRequest:
      type: object
      description: スタンプ更新リクエスト。
      properties:
        name:
          type: string
          description: スタンプ名
          example: "Gopher基礎"
          maxLength: 100

    UserStamp:
      type: object
      required:
        - user_id
        - stamp_id
      properties:
        user_id:
          type: integer
          format: int64
          description: ユーザーID
          example: 1
        stamp_id:
            type: integer
            format: int64
            description: スタンプID
            example: 1
        acquired_at:
          type: string
          format: date-time
          description: スタンプ取得日時
          example: "2023-01-01T00:00:00Z"

    AcquireStampRequest:
      type: object
      required:
        - stamp_id
        - otp
      properties:
        stamp_id:
          type: integer
          format: int64
          description: 取得するスタンプのID
          example: 1
        otp:
          type: string
          description: ワンタイムトークン（4桁の数字）
          example: "1234"

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: "INVALID_REQUEST"
        message:
          type: string
          description: エラーメッセージ
          example: "リクエストが不正です"
        details:
          type: string
          description: エラー詳細
          example: "name field is required"

tags:
  - name: Users
    description: ユーザー管理操作
  - name: Stamps
    description: スタンプマスターデータ管理操作
  - name: UserStamps
    description: ユーザーのスタンプ取得管理操作